model_loader_container/requirements.txt
transformers
torch
google-cloud-storage